requirements.txt
camelcase
charset-normalizer
Django
django-admin-decorators
django-admin-list-filter-dropdown
django-ckeditor-5
django-filter
django-js-asset
django-render-partial
django-humanize
djangorestframework
djangorestframework-simplejwt
fonttools
jsonfield
Khayyam
mysql-connector-python
Pillow
python-bsonjs
python-dateutil
python-dispatch
pytz
requests
humanize
django-jalali-date 
setuptools
PyJWT
virtualenv
django-jazzmin
crispy-bootstrap5
django-extensions
google-api-python-client
python-dotenv
django-meta
django-hcaptcha
django-environ
django-versatileimagefield
jdatetime
python-decouple
drf-yasg
drf-yasg[validation]
django-mail-templated